requirements.txt
requests
beautifulsoup4
pandas
openpyxl
textblob
vaderSentiment
python-dotenv
azure-storage-blob
